pyproject.toml
[tool.poetry]
name = "righttoask-directory"
version = "0.1.0"
description = ""
authors = ["Eleanor McMurtry <elem0@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
electionguard = "^1.1.15"
PyNaCl = "^1.4.0"
requests = "^2.25.1"
fastapi = "==0.63.0"
uvicorn = "==0.13.3"
uvloop = "==0.15.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
